La vista detalle ya funciona perfectamente, falta probar el escaner cuando se despliegue el servicio, en el emulador no deja, faltan favoritos, el buscador y las pruebas

+++ ServicioREST/REST_TFGMarioIngelmoDiana/src/main/java/es/unican/hapisecurity/REST_TFGMarioIngelmoDiana/controllerLayer/GeneralController.java
@@ -77,7 +77,7 @@ public class GeneralController {
 					.collect(Collectors.toList());
 		}
 		if (seguridad != null) {
-			dispositivos = dispositivos.stream().filter(d -> d.getSeguridad() >= Double.valueOf(seguridad))
+			dispositivos = dispositivos.stream().filter(d -> d.getSeguridad() >= Integer.valueOf(seguridad))
 					.collect(Collectors.toList());
 		}
 		if (sostenibilidad != null) {
